+++ src/Sortable.php
@@ -1,46 +1,15 @@
 <?php
+
 namespace Spatie\EloquentSortable;
 
 use Illuminate\Database\Query\Builder;
 
-trait Sortable
+interface Sortable
 {
     /**
      * Modify the order column value.
-     *
-     * @param $model
      */
-    public function setHighestOrderNumber($model)
-    {
-        $orderColumnName = $this->determineOrderColumnName();
-        $model->$orderColumnName = $this->getHighestOrderNumber();
-    }
-
-    /**
-     * Determine the column name of the order column.
-     *
-     * @return string
-     */
-    protected function determineOrderColumnName()
-    {
-        if (! isset($this->sortable['order_column_name']) || $this->sortable['order_column_name'] == '') {
-            $orderColumnName =  'order_column';
-        } else {
-            $orderColumnName = $this->sortable['order_column_name'];
-        }
-
-        return $orderColumnName;
-    }
-
-    /**
-     * Determine the order value for the new record.
-     *
-     * @return int
-     */
-    public function getHighestOrderNumber()
-    {
-        return ((int) self::max($this->determineOrderColumnName())) + 1;
-    }
+    public function setHighestOrderNumber();
 
     /**
      * Let's be nice and provide an ordered scope.
@@ -49,33 +18,15 @@ trait Sortable
      *
      * @return mixed
      */
-    public function scopeOrdered($query)
-    {
-        return $query->orderBy($this->determineOrderColumnName());
-    }
+    public function scopeOrdered($query);
 
     /**
      * This function reorders the records: the record with the first id in the array
      * will get order 1, the record with the second it will get order 2, ...
      *
-     * A starting order number can be optionally supplied (defaults to 1).
-     *
-     * @param array   $ids
-     * @param integer $startOrder
+     * @param array $ids
      *
      * @throws SortableException
      */
-    public static function setNewOrder($ids, $startOrder = 1)
-    {
-        if (! is_array($ids)) {
-            throw new SortableException('You must pass an array to setNewOrder');
-        }
-
-        foreach ($ids as $id) {
-            $model = self::find($id);
-            $orderColumnName = $model->determineOrderColumnName();
-            $model->$orderColumnName = $startOrder++;
-            $model->save();
-        }
-    }
+    public static function setNewOrder($ids);
 }