+++ src/SortableTrait.php
@@ -1,0 +1,81 @@
+<?php
+
+namespace Spatie\EloquentSortable;
+
+use Illuminate\Database\Query\Builder;
+
+trait SortableTrait
+{
+    /**
+     * Modify the order column value.
+     */
+    public function setHighestOrderNumber()
+    {
+        $orderColumnName = $this->determineOrderColumnName();
+        $this->$orderColumnName = $this->getHighestOrderNumber() + 1;
+    }
+
+    /**
+     * Determine the order value for the new record.
+     *
+     * @return int
+     */
+    public function getHighestOrderNumber()
+    {
+        return ((int) static::max($this->determineOrderColumnName()));
+    }
+
+    /**
+     * Let's be nice and provide an ordered scope.
+     *
+     * @param $query
+     *
+     * @return mixed
+     */
+    public function scopeOrdered($query)
+    {
+        return $query->orderBy($this->determineOrderColumnName());
+    }
+
+    /**
+     * This function reorders the records: the record with the first id in the array
+     * will get order 1, the record with the second it will get order 2, ...
+     *
+     * A starting order number can be optionally supplied (defaults to 1).
+     *
+     * @param array   $ids
+     * @param integer $startOrder
+     *
+     * @throws SortableException
+     */
+    public static function setNewOrder($ids, $startOrder = 1)
+    {
+        if (! is_array($ids)) {
+            throw new SortableException('You must pass an array to setNewOrder');
+        }
+
+        foreach ($ids as $id) {
+            $model = static::find($id);
+            $orderColumnName = $model->determineOrderColumnName();
+            $model->$orderColumnName = $startOrder++;
+            $model->save();
+        }
+    }
+
+    /**
+     * Determine the column name of the order column.
+     *
+     * @return string
+     */
+    protected function determineOrderColumnName()
+    {
+        if (
+            isset($this->sortable['order_column_name']) &&
+            !empty($this->sortable['order_column_name'])
+        ) {
+            return $this->sortable['order_column_name'];
+        }
+
+        return 'order_column';
+    }
+}