+++ tests/Dummy.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Spatie\EloquentSortable\Test;
+
+use Illuminate\Database\Eloquent\Model;
+use Spatie\EloquentSortable\Sortable;
+use Spatie\EloquentSortable\SortableTrait;
+
+class Dummy extends Model implements Sortable
+{
+    use SortableTrait;
+
+    protected $table = 'dummies';
+    protected $guarded = [];
+    public $timestamps = false;
+}