+++ tests/TestCase.php
@@ -1,0 +1,58 @@
+<?php
+
+namespace Spatie\EloquentSortable\Test;
+
+use Illuminate\Database\Schema\Blueprint;
+use Orchestra\Testbench\TestCase as Orchestra;
+
+abstract class TestCase extends Orchestra
+{
+    public function setUp()
+    {
+        parent::setUp();
+
+        $this->setUpDatabase();
+    }
+
+    /**
+     * @param \Illuminate\Foundation\Application $app
+     *
+     * @return array
+     */
+    protected function getPackageProviders($app)
+    {
+        return [
+            'Spatie\EloquentSortable\SortableServiceProvider',
+        ];
+    }
+
+    /**
+     * @param \Illuminate\Foundation\Application $app
+     * 
+     * @return void
+     */
+    protected function getEnvironmentSetUp($app)
+    {
+        $app['config']->set('database.default', 'sqlite');
+        $app['config']->set('database.connections.sqlite', [
+            'driver' => 'sqlite',
+            'database' => __DIR__.'/database.sqlite',
+            'prefix' => '',
+        ]);
+    }
+
+    protected function setUpDatabase()
+    {
+        file_put_contents(__DIR__.'/database.sqlite', null);
+
+        $this->app['db']->connection()->getSchemaBuilder()->create('dummies', function (Blueprint $table) {
+            $table->increments('id');
+            $table->string('name');
+            $table->integer('order_column');
+        });
+
+        for ($i = 1; $i <= 20; $i++) {
+            Dummy::create(['name' => $i]);
+        }
+    }
+}